+++ Cocoa/external_modules/code/PolyChordLite/.gitignore
@@ -1,0 +1,35 @@
+doc/html/
+doc/latex/
+*.mod
+*.swp
+*.swo
+*tags
+*.o
+*.dat
+main
+planck
+chains
+batch
+*.a
+*.so
+*.pyf
+getdist/getdist
+getdist/plot_data/*
+getdist/plots/*
+*.pyc
+bin/*
+*.gcda
+*.gcno
+*.gcov
+gmon.out
+core*
+*.egg-info
+build
+dist
+pypolychord/.ld_library_path.sh
+pypolychord/.ld_preload.sh
+.vscode
+*.pdf
+# Chains folder (test results only)
+chains/*
+.ipynb_checkpoints/
