+++ Cocoa/external_modules/code/PolyChordLite/Python_Functions/run_pypolychord_gaussian_shells.py
@@ -41,7 +41,7 @@ def likelihood(theta):
     r2 = 0
 
 
-    def logincexp(loga,logb,logx=None):
+    def logincexp(loga,logb,logc=None):
         if (loga>logb):
             loga = loga + log(exp(logb-loga) + 1)
         else:
