+++ Cocoa/external_modules/code/PolyChordLite/setup.py
@@ -92,7 +92,7 @@ class CustomBuildPy(_build_py, object):
             env["DEBUG"] = "1"
         
         BASE_PATH = os.path.dirname(os.path.abspath(__file__))
-        env["PWD"] = BASE_PATH
+        env["CURDIR"] = BASE_PATH
         env.update({k : os.environ[k] for k in ["CC", "CXX", "FC"] if k in os.environ})
         subprocess.check_call(["make", "-e", "libchord.so"], env=env, cwd=BASE_PATH)
         if not os.path.isdir("pypolychord/lib/"):
